separation of servlet and gen logic
